+++ example/flutter/install_flutter.dart
@@ -1,0 +1,7 @@
+import 'dart:async';
+
+import 'package:tekartik_build_utils/flutter/flutter.dart';
+
+Future main() async {
+  await installFlutter('.dart_tool/tekartik_build_utils/flutter');
+}