+++ lib/bin/dart_project_util.dart
@@ -6,7 +6,14 @@ import 'package:pub_semver/pub_semver.dart';
 import 'package:args/args.dart';
 
 final analysisOptionFilename = 'analysis_options.yaml';
-final commandGetAnalysisOption = 'analysis_options.yaml';
+final argGetAnalysisCommand = 'analysis_options.yaml';
+
+const String argTemplateOption = 'template';
+
+const simplePackageTemplate = 'simple_package';
+const String argImplicitCastsFlag = 'implicit-casts';
+const String noImplicitCastsTemplate = 'no_implicit_casts';
+const String implicitCastsTemplate = 'implicit_casts';
 
 const String argHelpFlag = 'help';
 const String argVersionFlag = 'version';
@@ -33,10 +40,15 @@ Future main(List<String> args) async {
   }
 
   var parser = ArgParser(allowTrailingOptions: false);
-  var analysisOptionParser = ArgParser();
+  var analysisOptionParser = ArgParser()
+    ..addFlag(argImplicitCastsFlag,
+        negatable: true, help: 'remove strong-mode: implicit-casts: false')
+    ..addOption(argTemplateOption, help: 'Template folder');
   _addHelp(analysisOptionParser);
-  parser.addCommand(commandGetAnalysisOption, analysisOptionParser);
+
+  parser.addCommand(argGetAnalysisCommand, analysisOptionParser);
   parser.addFlag(argDryRunFlag, abbr: 'd', help: "Don't execture the command");
+
   _addHelp(parser);
   parser.addFlag(argVersionFlag, help: 'Version', negatable: false);
   parser.addFlag(argVerboseFlag,
@@ -55,19 +67,19 @@ Future main(List<String> args) async {
     stdout.writeln(parser.commands.keys);
   }
 
-  bool version = result[argVersionFlag];
+  bool version = result[argVersionFlag] as bool;
   if (version) {
     _version();
     return;
   }
-  bool help = result[argHelpFlag];
+  bool help = result[argHelpFlag] as bool;
   if (help) {
     _help();
     return;
   }
 
   // devPrint(result.command?.name);
-  if (result.command?.name == commandGetAnalysisOption) {
+  if (result.command?.name == argGetAnalysisCommand) {
     var getAnalysisOptionResult = result.command;
 
     void _help() {
@@ -76,7 +88,7 @@ Future main(List<String> args) async {
       stdout.writeln(analysisOptionParser.usage);
     }
 
-    bool help = getAnalysisOptionResult[argHelpFlag];
+    bool help = getAnalysisOptionResult[argHelpFlag] as bool;
     if (help) {
       _help();
       return;
@@ -86,15 +98,23 @@ Future main(List<String> args) async {
       stdout.writeln(const LineSplitter().convert(text).first);
     }
 
+    String template;
+    if (getAnalysisOptionResult[argImplicitCastsFlag] == true) {
+      template = implicitCastsTemplate;
+    } else {
+      template = (getAnalysisOptionResult[argTemplateOption] as String) ??
+          noImplicitCastsTemplate;
+    }
     var urlRoot =
-        'https://raw.githubusercontent.com/tekartik/build_utils.dart/dart2/example/analyze/simple_package/analysis_options.yaml';
+        'https://raw.githubusercontent.com/tekartik/build_utils.dart/dart2/example/analyze';
+    var url = '${urlRoot}/${template}/${analysisOptionFilename}';
 
     var file = File(analysisOptionFilename);
     try {
       var oldYaml = await file.readAsString();
       _firstLine(oldYaml);
     } catch (_) {}
-    var analysisOptionsYaml = await read(urlRoot);
+    var analysisOptionsYaml = await read(url);
 
     await file.writeAsString(analysisOptionsYaml);
     _firstLine(analysisOptionsYaml);